change type of grade to floating point

+++ db/migrate/20170414191816_create_enrollments.rb
@@ -2,7 +2,7 @@ class CreateEnrollments < ActiveRecord::Migration[5.0]
   def change
     create_table :enrollments do |t|
 
-      t.integer :grade
+      t.float :grade
       
       t.references :section, null: false
       t.references :student, null: false
