+++ db/migrate/20170414192051_create_gpas.rb
@@ -5,7 +5,7 @@ class CreateGpas < ActiveRecord::Migration[5.0]
       t.integer :year, null: false
       t.integer :semester, null: false
       t.integer :credit, null: false
-      t.integer :gpa
+      t.float :gpa
       
       t.references :student, null: false
